requirements.txt
gradio
faiss-cpu
numpy
pandas
sentence-transformers
